Add file 06

+++ Hillel_HW_2_2.py
@@ -1,9 +1,9 @@
 
-# Варіант_1
-number_1 = float(input('Please enter a number_1 '))
-number_2 = float(input('Please enter a number_2 '))
-number_3 = float(input('Please enter a number_3 '))
+number_1 = int(input('Please enter a number_1 '))
+number_2 = int(input('Please enter a number_2 '))
+number_3 = int(input('Please enter a number_3 '))
 print((number_1 + number_2 + number_3) / 3)
+
 
 # Варіант_2_з_округленням
 # при числах менше 10, при округленні виходить 0
